Fixes #17

+++ eloquent/orm/model.py
@@ -46,6 +46,7 @@ class Model(object):
     __visible__ = []
 
     __timestamps__ = True
+    __dates__ = []
 
     __casts__ = {}
 
@@ -56,6 +57,7 @@ class Model(object):
     _with = []
 
     _booted = {}
+    _global_scopes = {}
     _registered = []
 
     __resolver = None
@@ -70,7 +72,6 @@ class Model(object):
         :param attributes: The instance attributes
         """
         self.__exists = False
-        self.__dates = []
         self.__original = {}
         self.__attributes = {}
         self.__relations = {}
@@ -97,7 +98,62 @@ class Model(object):
         """
         The booting method of the model.
         """
-        # TODO
+        # TODO: mutators
+
+        cls._boot_mixins()
+
+    @classmethod
+    def _boot_mixins(cls):
+        """
+        Boot the mixins
+        """
+        for mixin in cls.__bases__:
+            method = 'boot_%s' % inflection.underscore(mixin.__name__)
+            if hasattr(mixin, method):
+                getattr(mixin, method)(cls)
+
+    @classmethod
+    def add_global_scope(cls, scope):
+        """
+        Register a new global scope on the model.
+
+        :param scope: The scope to register
+        :type scope: eloquent.orm.scopes.scope.Scope
+        """
+        if cls not in cls._global_scopes:
+            cls._global_scopes[cls] = {}
+
+        cls._global_scopes[cls][scope.__class__] = scope
+
+    @classmethod
+    def has_global_scope(cls, scope):
+        """
+        Determine if a model has a global scope.
+
+        :param scope: The scope to register
+        :type scope: eloquent.orm.scopes.scope.Scope
+        """
+        return cls.get_global_scope(scope) is not None
+
+    @classmethod
+    def get_global_scope(cls, scope):
+        """
+        Get a global scope registered with the model.
+
+        :param scope: The scope to register
+        :type scope: eloquent.orm.scopes.scope.Scope
+        """
+        for key, value in cls._global_scopes[cls].items():
+            if isinstance(scope, key):
+                return value
+
+    def get_global_scopes(self):
+        """
+        Get the global scopes for this class instance.
+
+        :rtype: dict
+        """
+        return self.__class__._global_scopes.get(self.__class__, {})
 
     def fill(self, **attributes):
         """
@@ -896,6 +952,9 @@ class Model(object):
         """
         Perform the actual delete query on this model instance.
         """
+        if hasattr(self, '_do_perform_delete_on_model'):
+            return self._do_perform_delete_on_model()
+
         return self.new_query().where(self.get_key_name(), self.get_key()).delete()
 
     # TODO: events
@@ -1237,11 +1296,62 @@ class Model(object):
         :return: A Builder instance
         :rtype: Builder
         """
+        builder = self.new_query_without_scopes()
+
+        return self.apply_global_scopes(builder)
+
+    def new_query_without_scope(self, scope):
+        """
+        Get a new query builder for the model's table without a given scope
+
+        :return: A Builder instance
+        :rtype: Builder
+        """
+        builder = self.new_query()
+        self.get_global_scope(scope).remove(builder, self)
+
+        return builder
+
+    def new_query_without_scopes(self):
+        """
+        Get a new query builder without any scopes.
+
+        :return: A Builder instance
+        :rtype: Builder
+        """
         builder = self.new_orm_builder(
             self._new_base_query_builder()
         )
 
         return builder.set_model(self).with_(*self._with)
+
+    def apply_global_scopes(self, builder):
+        """
+        Apply all of the global scopes to a builder.
+
+        :param builder: A Builder instance
+        :type builder: Builder
+
+        :rtype: Builder
+        """
+        for scope in self.get_global_scopes().values():
+            scope.apply(builder, self)
+
+        return builder
+
+    def remove_global_scopes(self, builder):
+        """
+        Remove all of the global scopes from a builder.
+
+        :param builder: A Builder instance
+        :type builder: Builder
+
+        :rtype: Builder
+        """
+        for scope in self.get_global_scopes().values():
+            scope.remove(builder, self)
+
+        return builder
 
     @classmethod
     def query(cls):
@@ -1827,7 +1937,7 @@ class Model(object):
         """
         defaults = [self.CREATED_AT, self.UPDATED_AT]
 
-        return self.__dates + defaults
+        return self.__dates__ + defaults
 
     def from_datetime(self, value):
         """
